Crate storage_interface

Modules

async_proof_fetcher

cached_state_view

proof_fetcher

state_delta

state_view

sync_proof_fetcher

Structs

DbReaderWriter

ExecutedTrees

A wrapper of the in-memory state sparse merkle tree and the transaction accumulator that represent a specific state collectively. Usually it is a state after executing a block.

GetStateValueByVersionRequest

SaveTransactionsRequest

Enums

Error

Order

StorageRequest

Network types for storage service

Traits

DbReader

Trait that is implemented by a DB that supports certain public (to client) read APIs expected of an Aptos DB

DbWriter

Trait that is implemented by a DB that supports certain public (to client) write APIs expected of an Aptos DB. This adds write APIs to DbReader.

StateSnapshotReceiver

Functions

jmt_update_refs

jmt_updates