Module storage_interface::cached_state_view

Structs

CachedDbStateView

CachedStateView

CachedStateView is like a snapshot of the global state comprised of state view at two levels, persistent storage and memory.

StateCache