Trait storage_interface::proof_fetcher::ProofFetcher

pub trait ProofFetcher: Sync + Send {
    fn fetch_state_value_and_proof(
        &self, 
        state_key: &StateKey, 
        version: Version, 
        root_hash: Option<HashValue>
    ) -> Result<(Option<StateValue>, Option<SparseMerkleProofExt>)>;
    fn get_proof_cache(&self) -> HashMap<HashValue, SparseMerkleProofExt>;
}

Defines the trait for fetching proof from the DB

Required Methods

fn fetch_state_value_and_proof(
&self,
state_key: &StateKey,
version: Version,
root_hash: Option<HashValue>
) -> Result<(Option<StateValue>, Option<SparseMerkleProofExt>)>

Fetches the state value along with its proof. If root_hash is provided, will also verify its proof.

fn get_proof_cache(&self) -> HashMap<HashValue, SparseMerkleProofExt>

API to return all the proofs fetched by the proof fetcher so far.