Trait storage_interface::DbWriter

pub trait DbWriter: Send + Sync {
    fn get_state_snapshot_receiver(
        &self, 
        version: Version, 
        expected_root_hash: HashValue
    ) -> Result<Box<dyn StateSnapshotReceiver<StateKey, StateValue>>> { ... }
    fn finalize_state_snapshot(
        &self, 
        version: Version, 
        output_with_proof: TransactionOutputListWithProof, 
        ledger_infos: &[LedgerInfoWithSignatures]
    ) -> Result<()> { ... }
    fn save_transactions(
        &self, 
        txns_to_commit: &[TransactionToCommit], 
        first_version: Version, 
        base_state_version: Option<Version>, 
        ledger_info_with_sigs: Option<&LedgerInfoWithSignatures>, 
        sync_commit: bool, 
        latest_in_memory_state: StateDelta
    ) -> Result<()> { ... }
}

Trait that is implemented by a DB that supports certain public (to client) write APIs expected of an Aptos DB. This adds write APIs to DbReader.

Provided Methods

fn get_state_snapshot_receiver(
&self,
version: Version,
expected_root_hash: HashValue
) -> Result<Box<dyn StateSnapshotReceiver<StateKey, StateValue>>>

Get a (stateful) state snapshot receiver.

Chunk of accounts need to be added via add_chunk() before finishing up with finish_box()

fn finalize_state_snapshot(
&self,
version: Version,
output_with_proof: TransactionOutputListWithProof,
ledger_infos: &[LedgerInfoWithSignatures]
) -> Result<()>

Finalizes a state snapshot that has already been restored to the database through a state snapshot receiver. This is required to bootstrap the transaction accumulator, populate transaction information, save the epoch ending ledger infos and delete genesis.

Note: this assumes that the output with proof has already been verified and that the state snapshot was restored at the same version.

fn save_transactions(
&self,
txns_to_commit: &[TransactionToCommit],
first_version: Version,
base_state_version: Option<Version>,
ledger_info_with_sigs: Option<&LedgerInfoWithSignatures>,
sync_commit: bool,
latest_in_memory_state: StateDelta
) -> Result<()>

Persist transactions. Called by the executor module when either syncing nodes or committing blocks during normal operation. See AptosDB::save_transactions.